Server/requirements.txt
numpy
pandas
matplotlib
scikit-learn
imblearn
torch
torchvision
accelerate
evaluate
datasets>=2.14.5
transformers
tqdm
Pillow
huggingface_hub
transformers==4.32.0